Fix some long to int conversion warnings.

+++ src/Core/Entities/PlayerEntity.h
@@ -542,7 +542,7 @@ typedef enum
 	// target memory
 	// TODO: this should use weakrefs
 	NSMutableArray  		*target_memory;
-	int						target_memory_index;
+	NSInteger				target_memory_index;
 	
 	// custom view points
 	Quaternion				customViewQuaternion;
